+++ aiosmtpd/docs/NEWS.rst
@@ -15,7 +15,7 @@ Fixed/Improved
 * All Controllers now have more rationale design, as they are now composited from a Base + a Mixin
 
 
-1.4.2 (aiosmtpd-next)
+1.4.2 (2021-03-08)
 =====================
 
 Fixed/Improved
@@ -28,7 +28,8 @@ Fixed/Improved
   (See #262)
 * Timeout messages in ``Controller.start()`` gets more details and a mention about the
   ``ready_timeout`` parameter. (See #262)
-
+* Prevent sensitive AUTH information leak by sanitizing the repr()
+  of AuthResult and LoginPassword.
 
 
 1.4.1 (2021-03-04)
@@ -37,7 +38,6 @@ Fixed/Improved
 Fixed/Improved
 --------------
 * Maximum length of email address local part is customizable, defaults to no limit. (Closes #257)
-
 
 
 1.4.0 (2021-02-26)